[Cherry-Pick] 将Converters提取为Class
