+++ HyPlayer/Controls/SongsList.xaml.cs
@@ -410,18 +410,3 @@ public sealed partial class SongsList : UserControl, IDisposable
         _ = IndicateNowPlayingItem();
     }
 }
-
-public class SongListSelectModeConverter : IValueConverter
-{
-    public object Convert(object value, Type targetType, object parameter, string language)
-    {
-        if (value != null)
-            return (bool)value ? ListViewSelectionMode.Multiple : ListViewSelectionMode.Single;
-        return SelectionMode.Single;
-    }
-
-    public object ConvertBack(object value, Type targetType, object parameter, string language)
-    {
-        throw new NotImplementedException();
-    }
-}
