+++ HyPlayer/MainPage.xaml.cs
@@ -66,29 +66,3 @@ public sealed partial class MainPage
         }
     }
 }
-
-public class PlayBarMarginConverter : IValueConverter
-{
-    public object Convert(object value, Type targetType, object parameter, string language)
-    {
-        return value is true ? new Thickness(16) : new Thickness(0);
-    }
-
-    public object ConvertBack(object value, Type targetType, object parameter, string language)
-    {
-        throw new NotImplementedException();
-    }
-}
-
-public class PlayBarCornerRadiusConverter : IValueConverter
-{
-    public object Convert(object value, Type targetType, object parameter, string language)
-    {
-        return value is true ? new CornerRadius(8) : new CornerRadius(8, 8, 0, 0);
-    }
-
-    public object ConvertBack(object value, Type targetType, object parameter, string language)
-    {
-        throw new NotImplementedException();
-    }
-}
