+++ HyPlayer/Pages/Settings.xaml.cs
@@ -411,16 +411,3 @@ public sealed partial class Settings : Page
         LastFMManager.TryLogoffLastFM();
     }
 }
-public class ReversedBooleanToVisibilityConverter : IValueConverter
-{
-    public object Convert(object value, Type targetType, object parameter, string language)
-    {
-        if (value is true) return Visibility.Collapsed;
-        return Visibility.Visible;
-    }
-
-    public object ConvertBack(object value, Type targetType, object parameter, string language)
-    {
-        throw new NotImplementedException();
-    }
-}
