+++ HyPlayer/Resources/Acrylic.xaml
@@ -1,7 +1,8 @@
 ﻿<ResourceDictionary
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:media="http://schemas.microsoft.com/winfx/2006/xaml/presentation">
+    xmlns:media="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:local="using:HyPlayer">
     <media:AcrylicBrush
         x:Key="PlaybackAcrylicInAppLuminosity"
         BackgroundSource="Backdrop"
@@ -44,4 +45,18 @@
         TintLuminosityOpacity="0.36"
         TintColor="#4D76B3"
         FallbackColor="#2B2B2B" />
+    <AcrylicBrush 
+        x:Key="TransparentBackgroundAcrylic"
+        BackgroundSource="HostBackdrop"
+        TintColor="{ThemeResource SystemRevealAltHighColor}"
+        TintOpacity="1"
+        TintLuminosityOpacity="0.5"
+        FallbackColor="{ThemeResource SystemRevealAltHighColor}"/>
+    <AcrylicBrush 
+        x:Key="NormalBackgroundAcrylic"
+        BackgroundSource="HostBackdrop"
+        TintColor="{ThemeResource SystemRevealAltHighColor}"
+        TintOpacity="0.5"
+        TintLuminosityOpacity="0.5"
+        FallbackColor="{ThemeResource SystemRevealAltHighColor}"/>
 </ResourceDictionary>
